Possible solution to multiple sources of truth
